Move `EdgeEndpoint` to `subgraph` module
